backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
asyncpg
psycopg2-binary
alembic
passlib
python-jose
bcrypt==4.0.1
python-multipart